backend/requirements.txt
boto3==1.26.91
Django==4.1.7
django_environ==0.10.0
django_filter==22.1
django_on_heroku==1.1.2
djangorestframework==3.14.0
djoser==2.2.0
environ==1.0
itemadapter==0.7.0
python_scrapyd_api==2.1.2
requests==2.24.0
Scrapy==2.8.0
scrapyd_api==0.0.1
